BugFix: keep count of added proxies in `added_proxies` before commiting the changes
the bug was that when it tried to exclude inbounds, it was looking up the database and checking if there is any proxy that needed inbound exclusion.
the problem was when a proxy was just added and wasn't commited to the database yet, so it was thinking that the proxy didn't exist.