Record_node_usages hang due to lack of correct timeout handling in streaming
https://github.com/Gozargah/Marzban/issues/1550#issue-2762388346
When the record_node_usages task is executed, a stream gRPC call to QueryStats is made. If during the stream iteration the waiting time is exceeded. grpc.RpcError exception with DEADLINE_EXCEEDED code occurs. Since this exception is not handled correctly, the generator hangs and the task does not complete in time. As a result, APScheduler cannot start a new task instance (due to max_instances=1), which leads to accumulation of ‘hung’ tasks.
SOLUTION
Add grpc.RpcError exception handling with DEADLINE_EXCEEDED code both during stream iteration and outside the loop. When such an exception occurs, the generator should terminate correctly.